If checked series only with outside value between average value and series value from dopdown list choosen can show

+++ web/controller/app.go
@@ -66,7 +66,7 @@ type PayloadAnalyticTLP struct {
 	DateEnd   	time.Time
 	ColName 	string
 	DeviationStatus  	bool
-	Deviation  	int
+	Deviation  	float64
 }
